useEffect => Listening to a global browser event (pointermove)
In this example, the external system is the browser DOM itself. Normally, you’d specify event listeners with JSX, but you can’t listen to the global window object this way. An Effect lets you connect to the window object and listen to its events. Listening to the pointermove event lets you track the cursor (or finger) position and update the red dot to move with it.